requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
pydantic==2.5.0
pydantic-settings==2.1.0

# ML & Embeddings
numpy==1.24.4
scikit-learn==1.3.2
sentence-transformers==2.2.2
huggingface_hub==0.25.2  # Add this line for compatibility
faiss-cpu==1.7.4


# Document Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
python-magic-bin==0.4.14
beautifulsoup4==4.12.2

# Google Gemini
google-generativeai==0.3.2

# Database
aiosqlite==0.19.0
asyncpg==0.29.0

# Caching & Performance
redis==5.0.1
python-redis-lock==4.0.0
cachetools==5.3.2

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1